requirements.txt
pandas == 1.5.3
pillow
requests
numpy
tensorflow
scikit-learn
matplotlib